Move config class out of init
